fix: resolve most if not all undefined size/sign conversions

+++ glass/src/cpu/gdt/gdt.h
@@ -50,7 +50,7 @@ typedef struct {
     uintptr_t base;       
 } __attribute__((packed)) gdtr_t;
 
-void        gdt_add_descriptor(uint64_t base, uint64_t limit, uint8_t access, uint8_t granularity);
+void        gdt_add_descriptor(uint64_t base, uint16_t limit, uint8_t access, uint8_t granularity);
 void        gdt_reload(gdtr_t* gdtr, uint16_t code, uint16_t data);
 uint16_t    gdt_install_tss(uint64_t tss);
 void        gdt_assemble();
